Add editable updates
Closes #30
