+++ components/updates/server-actions.ts
@@ -49,6 +49,9 @@ export const upsertUpdate = async (update: Partial<Update>) => {
     data: { user },
   } = await supabase.auth.getUser();
 
+  // Only include posted_by if this is a new update (no id)
+  const posted_by = update.id ? undefined : user?.id || null;
+
   const { data, error } = await supabase
     .from('updates')
     .upsert({
@@ -60,28 +63,27 @@ export const upsertUpdate = async (update: Partial<Update>) => {
       description: update.description,
       value: update.value || null,
       link: update.link || null,
-      posted_by: user?.id || null,
+      posted_by,
       year: update.year || new Date().getFullYear(),
       impact_indicator_id: update.impact_indicator_id,
-      source: 'Maerl',
-      original: update.original || true,
-      duplicate: update.duplicate || false,
-      verified: update.verified || false,
-      valid: update.valid || true,
+      source: update.source || 'Maerl',
+      original: update.original ?? true,
+      duplicate: update.duplicate ?? false,
+      verified: update.verified ?? false,
+      valid: update.valid ?? true,
     })
     .select()
     .single();
 
   if (error) throw error;
 
-  const { data: updatedProject } = await supabase
+  // Update the project's last_updated timestamp
+  await supabase
     .from('projects')
     .update({ last_updated: new Date().toISOString() })
-    .eq('id', update.project_id)
-    .select()
-    .single();
+    .eq('id', update.project_id);
 
-  return { update };
+  return { update: data };
 };
 
 export const fetchOutputUpdates = async (outputId: string) => {
