--- components/updates/update-editing-form.tsx
@@ -1,9 +1,0 @@
-import { Update } from '@/utils/types';
-
-export default function UpdateEditingForm({ update }: { update: Update }) {
-  return (
-    <div>
-      <p>{update.description}</p>
-    </div>
-  );
-}
